фикс в стилях и работе приложения
