+++ index.html
@@ -438,7 +438,7 @@
                                     </div>
                                     <div class="track-play__dislike _btn-icon">
                                         <svg class="track-play__dislike-svg" alt="dislike">
-                                            <use xlink:href="img/icon/sprite.svg#icon-dislike"></use>
+                                            <use xlink:href="public/img/icon/sprite.svg#icon-dislike"></use>
                                         </svg>
                                     </div>
                                 </div>
